Make logical volume groups immutable
